Implement data-model-v4 changes into the model classes
