+++ src/main/java/co/edu/icesi/researchgroupmanagement/model/AdmissionRequest.java
@@ -1,14 +1,11 @@
 package co.edu.icesi.researchgroupmanagement.model;
-
-// Generated Oct 15, 2022, 11:56:22 PM by Hibernate Tools 5.6.7.Final
-
-import java.util.Date;
+// Generated Oct 22, 2022, 7:23:43 PM by Hibernate Tools 5.6.7.Final
 
 import javax.persistence.Entity;
+import javax.persistence.Id;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
+import java.util.Date;
 
 /**
  * AdmissionRequest generated by hbm2java
