+++ src/main/java/co/edu/icesi/researchgroupmanagement/model/AdmissionStatus.java
@@ -1,0 +1,61 @@
+package co.edu.icesi.researchgroupmanagement.model;
+// Generated Oct 22, 2022, 7:23:43 PM by Hibernate Tools 5.6.7.Final
+
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import java.util.HashSet;
+import java.util.Set;
+
+/**
+ * AdmissionStatus generated by hbm2java
+ */
+@Entity
+public class AdmissionStatus implements java.io.Serializable {
+
+	@Id
+	@GeneratedValue(strategy= GenerationType.SEQUENCE)
+	private long id;
+	private String status;
+	private Set admissionRequests = new HashSet(0);
+
+	public AdmissionStatus() {
+	}
+
+	public AdmissionStatus(long id, String status) {
+		this.id = id;
+		this.status = status;
+	}
+
+	public AdmissionStatus(long id, String status, Set admissionRequests) {
+		this.id = id;
+		this.status = status;
+		this.admissionRequests = admissionRequests;
+	}
+
+	public long getId() {
+		return this.id;
+	}
+
+	public void setId(long id) {
+		this.id = id;
+	}
+
+	public String getStatus() {
+		return this.status;
+	}
+
+	public void setStatus(String status) {
+		this.status = status;
+	}
+
+	public Set getAdmissionRequests() {
+		return this.admissionRequests;
+	}
+
+	public void setAdmissionRequests(Set admissionRequests) {
+		this.admissionRequests = admissionRequests;
+	}
+
+}
