+++ src/main/java/co/edu/icesi/researchgroupmanagement/model/Faculty.java
@@ -1,41 +1,36 @@
 package co.edu.icesi.researchgroupmanagement.model;
+// Generated Oct 22, 2022, 7:23:43 PM by Hibernate Tools 5.6.7.Final
 
-// Generated Oct 15, 2022, 11:56:22 PM by Hibernate Tools 5.6.7.Final
-
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
 import java.util.HashSet;
 import java.util.Set;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
 /**
- * SciTiNationalProgram generated by hbm2java
+ * Faculty generated by hbm2java
  */
 @Entity
-public class SciTiNationalProgram implements java.io.Serializable {
+public class Faculty implements java.io.Serializable {
 
 	@Id
 	@GeneratedValue(strategy= GenerationType.SEQUENCE)
 	private long id;
 	private String name;
-	private String description;
 	private Set researchGroups = new HashSet(0);
 
-	public SciTiNationalProgram() {
+	public Faculty() {
 	}
 
-	public SciTiNationalProgram(long id, String name, String description) {
+	public Faculty(long id, String name) {
 		this.id = id;
 		this.name = name;
-		this.description = description;
 	}
 
-	public SciTiNationalProgram(long id, String name, String description, Set researchGroups) {
+	public Faculty(long id, String name, Set researchGroups) {
 		this.id = id;
 		this.name = name;
-		this.description = description;
 		this.researchGroups = researchGroups;
 	}
 
@@ -53,14 +48,6 @@ public class SciTiNationalProgram implements java.io.Serializable {
 
 	public void setName(String name) {
 		this.name = name;
-	}
-
-	public String getDescription() {
-		return this.description;
-	}
-
-	public void setDescription(String description) {
-		this.description = description;
 	}
 
 	public Set getResearchGroups() {
