+++ src/main/java/co/edu/icesi/researchgroupmanagement/model/FinancingSource.java
@@ -1,38 +1,36 @@
 package co.edu.icesi.researchgroupmanagement.model;
+// Generated Oct 22, 2022, 7:23:43 PM by Hibernate Tools 5.6.7.Final
 
-// Generated Oct 15, 2022, 11:56:22 PM by Hibernate Tools 5.6.7.Final
-
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
 import java.util.HashSet;
 import java.util.Set;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
 /**
- * RprojectTypes generated by hbm2java
+ * FinancingSource generated by hbm2java
  */
 @Entity
-public class RprojectTypes implements java.io.Serializable {
+public class FinancingSource implements java.io.Serializable {
 
 	@Id
 	@GeneratedValue(strategy= GenerationType.SEQUENCE)
 	private long id;
-	private String type;
+	private String source;
 	private Set researchProjects = new HashSet(0);
 
-	public RprojectTypes() {
+	public FinancingSource() {
 	}
 
-	public RprojectTypes(long id, String type) {
+	public FinancingSource(long id, String source) {
 		this.id = id;
-		this.type = type;
+		this.source = source;
 	}
 
-	public RprojectTypes(long id, String type, Set researchProjects) {
+	public FinancingSource(long id, String source, Set researchProjects) {
 		this.id = id;
-		this.type = type;
+		this.source = source;
 		this.researchProjects = researchProjects;
 	}
 
@@ -44,12 +42,12 @@ public class RprojectTypes implements java.io.Serializable {
 		this.id = id;
 	}
 
-	public String getType() {
-		return this.type;
+	public String getSource() {
+		return this.source;
 	}
 
-	public void setType(String type) {
-		this.type = type;
+	public void setSource(String source) {
+		this.source = source;
 	}
 
 	public Set getResearchProjects() {
