+++ src/main/java/co/edu/icesi/researchgroupmanagement/model/FinancingType.java
@@ -1,38 +1,36 @@
 package co.edu.icesi.researchgroupmanagement.model;
+// Generated Oct 22, 2022, 7:23:43 PM by Hibernate Tools 5.6.7.Final
 
-// Generated Oct 15, 2022, 11:56:22 PM by Hibernate Tools 5.6.7.Final
-
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
 import java.util.HashSet;
 import java.util.Set;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
 /**
- * RprojectStatus generated by hbm2java
+ * FinancingType generated by hbm2java
  */
 @Entity
-public class RprojectStatus implements java.io.Serializable {
+public class FinancingType implements java.io.Serializable {
 
 	@Id
 	@GeneratedValue(strategy= GenerationType.SEQUENCE)
 	private long id;
-	private String status;
+	private String type;
 	private Set researchProjects = new HashSet(0);
 
-	public RprojectStatus() {
+	public FinancingType() {
 	}
 
-	public RprojectStatus(long id, String status) {
+	public FinancingType(long id, String type) {
 		this.id = id;
-		this.status = status;
+		this.type = type;
 	}
 
-	public RprojectStatus(long id, String status, Set researchProjects) {
+	public FinancingType(long id, String type, Set researchProjects) {
 		this.id = id;
-		this.status = status;
+		this.type = type;
 		this.researchProjects = researchProjects;
 	}
 
@@ -44,12 +42,12 @@ public class RprojectStatus implements java.io.Serializable {
 		this.id = id;
 	}
 
-	public String getStatus() {
-		return this.status;
+	public String getType() {
+		return this.type;
 	}
 
-	public void setStatus(String status) {
-		this.status = status;
+	public void setType(String type) {
+		this.type = type;
 	}
 
 	public Set getResearchProjects() {
