+++ src/main/java/co/edu/icesi/researchgroupmanagement/model/MincienciasCategory.java
@@ -1,38 +1,36 @@
 package co.edu.icesi.researchgroupmanagement.model;
+// Generated Oct 22, 2022, 7:23:43 PM by Hibernate Tools 5.6.7.Final
 
-// Generated Oct 15, 2022, 11:56:22 PM by Hibernate Tools 5.6.7.Final
-
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
 import java.util.HashSet;
 import java.util.Set;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
 /**
- * Faculty generated by hbm2java
+ * MincienciasCategory generated by hbm2java
  */
 @Entity
-public class Faculty implements java.io.Serializable {
+public class MincienciasCategory implements java.io.Serializable {
 
 	@Id
 	@GeneratedValue(strategy= GenerationType.SEQUENCE)
 	private long id;
-	private String name;
+	private String category;
 	private Set researchGroups = new HashSet(0);
 
-	public Faculty() {
+	public MincienciasCategory() {
 	}
 
-	public Faculty(long id, String name) {
+	public MincienciasCategory(long id, String category) {
 		this.id = id;
-		this.name = name;
+		this.category = category;
 	}
 
-	public Faculty(long id, String name, Set researchGroups) {
+	public MincienciasCategory(long id, String category, Set researchGroups) {
 		this.id = id;
-		this.name = name;
+		this.category = category;
 		this.researchGroups = researchGroups;
 	}
 
@@ -44,12 +42,12 @@ public class Faculty implements java.io.Serializable {
 		this.id = id;
 	}
 
-	public String getName() {
-		return this.name;
+	public String getCategory() {
+		return this.category;
 	}
 
-	public void setName(String name) {
-		this.name = name;
+	public void setCategory(String category) {
+		this.category = category;
 	}
 
 	public Set getResearchGroups() {
