+++ src/main/java/co/edu/icesi/researchgroupmanagement/model/ResearchGroup.java
@@ -1,14 +1,12 @@
 package co.edu.icesi.researchgroupmanagement.model;
-
-// Generated Oct 15, 2022, 11:56:22 PM by Hibernate Tools 5.6.7.Final
-
-import java.util.HashSet;
-import java.util.Set;
+// Generated Oct 22, 2022, 7:23:43 PM by Hibernate Tools 5.6.7.Final
 
 import javax.persistence.Entity;
+import javax.persistence.Id;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
-import javax.persistence.Id;
+import java.util.HashSet;
+import java.util.Set;
 
 /**
  * ResearchGroup generated by hbm2java
@@ -21,54 +19,43 @@ public class ResearchGroup implements java.io.Serializable {
 	private long id;
 	private Faculty faculty;
 	private Institution institution;
+	private MincienciasCategory mincienciasCategory;
 	private SciTiNationalProgram sciTiNationalProgram;
-	private Users users;
 	private String name;
-	private String colcienciasCategory;
 	private String description;
-	private String citizenId;
-	private String nationality;
-	private String secondLastName;
-	private String countryOfBirth;
 	private Set admissionRequests = new HashSet(0);
 	private Set rgroupUserses = new HashSet(0);
 	private Set researchLines = new HashSet(0);
+	private Set rgroupLeaderses = new HashSet(0);
 	private Set researchProjects = new HashSet(0);
 
 	public ResearchGroup() {
 	}
 
-	public ResearchGroup(long id, Users users, String name, String colcienciasCategory, String description,
-			String citizenId, String nationality, String countryOfBirth) {
-		this.id = id;
-		this.users = users;
-		this.name = name;
-		this.colcienciasCategory = colcienciasCategory;
-		this.description = description;
-		this.citizenId = citizenId;
-		this.nationality = nationality;
-		this.countryOfBirth = countryOfBirth;
-	}
-
-	public ResearchGroup(long id, Faculty faculty, Institution institution, SciTiNationalProgram sciTiNationalProgram,
-			Users users, String name, String colcienciasCategory, String description, String citizenId,
-			String nationality, String secondLastName, String countryOfBirth, Set admissionRequests, Set rgroupUserses,
-			Set researchLines, Set researchProjects) {
+	public ResearchGroup(long id, Faculty faculty, Institution institution, MincienciasCategory mincienciasCategory,
+			String name, String description) {
 		this.id = id;
 		this.faculty = faculty;
 		this.institution = institution;
-		this.sciTiNationalProgram = sciTiNationalProgram;
-		this.users = users;
+		this.mincienciasCategory = mincienciasCategory;
 		this.name = name;
-		this.colcienciasCategory = colcienciasCategory;
 		this.description = description;
-		this.citizenId = citizenId;
-		this.nationality = nationality;
-		this.secondLastName = secondLastName;
-		this.countryOfBirth = countryOfBirth;
+	}
+
+	public ResearchGroup(long id, Faculty faculty, Institution institution, MincienciasCategory mincienciasCategory,
+			SciTiNationalProgram sciTiNationalProgram, String name, String description, Set admissionRequests,
+			Set rgroupUserses, Set researchLines, Set rgroupLeaderses, Set researchProjects) {
+		this.id = id;
+		this.faculty = faculty;
+		this.institution = institution;
+		this.mincienciasCategory = mincienciasCategory;
+		this.sciTiNationalProgram = sciTiNationalProgram;
+		this.name = name;
+		this.description = description;
 		this.admissionRequests = admissionRequests;
 		this.rgroupUserses = rgroupUserses;
 		this.researchLines = researchLines;
+		this.rgroupLeaderses = rgroupLeaderses;
 		this.researchProjects = researchProjects;
 	}
 
@@ -96,20 +83,20 @@ public class ResearchGroup implements java.io.Serializable {
 		this.institution = institution;
 	}
 
+	public MincienciasCategory getMincienciasCategory() {
+		return this.mincienciasCategory;
+	}
+
+	public void setMincienciasCategory(MincienciasCategory mincienciasCategory) {
+		this.mincienciasCategory = mincienciasCategory;
+	}
+
 	public SciTiNationalProgram getSciTiNationalProgram() {
 		return this.sciTiNationalProgram;
 	}
 
 	public void setSciTiNationalProgram(SciTiNationalProgram sciTiNationalProgram) {
 		this.sciTiNationalProgram = sciTiNationalProgram;
-	}
-
-	public Users getUsers() {
-		return this.users;
-	}
-
-	public void setUsers(Users users) {
-		this.users = users;
 	}
 
 	public String getName() {
@@ -120,52 +107,12 @@ public class ResearchGroup implements java.io.Serializable {
 		this.name = name;
 	}
 
-	public String getColcienciasCategory() {
-		return this.colcienciasCategory;
-	}
-
-	public void setColcienciasCategory(String colcienciasCategory) {
-		this.colcienciasCategory = colcienciasCategory;
-	}
-
 	public String getDescription() {
 		return this.description;
 	}
 
 	public void setDescription(String description) {
 		this.description = description;
-	}
-
-	public String getCitizenId() {
-		return this.citizenId;
-	}
-
-	public void setCitizenId(String citizenId) {
-		this.citizenId = citizenId;
-	}
-
-	public String getNationality() {
-		return this.nationality;
-	}
-
-	public void setNationality(String nationality) {
-		this.nationality = nationality;
-	}
-
-	public String getSecondLastName() {
-		return this.secondLastName;
-	}
-
-	public void setSecondLastName(String secondLastName) {
-		this.secondLastName = secondLastName;
-	}
-
-	public String getCountryOfBirth() {
-		return this.countryOfBirth;
-	}
-
-	public void setCountryOfBirth(String countryOfBirth) {
-		this.countryOfBirth = countryOfBirth;
 	}
 
 	public Set getAdmissionRequests() {
@@ -190,6 +137,14 @@ public class ResearchGroup implements java.io.Serializable {
 
 	public void setResearchLines(Set researchLines) {
 		this.researchLines = researchLines;
+	}
+
+	public Set getRgroupLeaderses() {
+		return this.rgroupLeaderses;
+	}
+
+	public void setRgroupLeaderses(Set rgroupLeaderses) {
+		this.rgroupLeaderses = rgroupLeaderses;
 	}
 
 	public Set getResearchProjects() {
