+++ src/main/java/co/edu/icesi/researchgroupmanagement/model/RgroupLeaders.java
@@ -1,0 +1,105 @@
+package co.edu.icesi.researchgroupmanagement.model;
+// Generated Oct 22, 2022, 7:23:43 PM by Hibernate Tools 5.6.7.Final
+
+/**
+ * RgroupLeaders generated by hbm2java
+ */
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
+@Entity
+public class RgroupLeaders implements java.io.Serializable {
+
+	@Id
+	@GeneratedValue(strategy= GenerationType.SEQUENCE)
+	private RgroupLeadersId id;
+	private ResearchGroup researchGroup;
+	private Users users;
+	private String citizenId;
+	private String nationality;
+	private String secondLastName;
+	private String countryOfBirth;
+
+	public RgroupLeaders() {
+	}
+
+	public RgroupLeaders(RgroupLeadersId id, ResearchGroup researchGroup, Users users, String citizenId,
+			String nationality, String countryOfBirth) {
+		this.id = id;
+		this.researchGroup = researchGroup;
+		this.users = users;
+		this.citizenId = citizenId;
+		this.nationality = nationality;
+		this.countryOfBirth = countryOfBirth;
+	}
+
+	public RgroupLeaders(RgroupLeadersId id, ResearchGroup researchGroup, Users users, String citizenId,
+			String nationality, String secondLastName, String countryOfBirth) {
+		this.id = id;
+		this.researchGroup = researchGroup;
+		this.users = users;
+		this.citizenId = citizenId;
+		this.nationality = nationality;
+		this.secondLastName = secondLastName;
+		this.countryOfBirth = countryOfBirth;
+	}
+
+	public RgroupLeadersId getId() {
+		return this.id;
+	}
+
+	public void setId(RgroupLeadersId id) {
+		this.id = id;
+	}
+
+	public ResearchGroup getResearchGroup() {
+		return this.researchGroup;
+	}
+
+	public void setResearchGroup(ResearchGroup researchGroup) {
+		this.researchGroup = researchGroup;
+	}
+
+	public Users getUsers() {
+		return this.users;
+	}
+
+	public void setUsers(Users users) {
+		this.users = users;
+	}
+
+	public String getCitizenId() {
+		return this.citizenId;
+	}
+
+	public void setCitizenId(String citizenId) {
+		this.citizenId = citizenId;
+	}
+
+	public String getNationality() {
+		return this.nationality;
+	}
+
+	public void setNationality(String nationality) {
+		this.nationality = nationality;
+	}
+
+	public String getSecondLastName() {
+		return this.secondLastName;
+	}
+
+	public void setSecondLastName(String secondLastName) {
+		this.secondLastName = secondLastName;
+	}
+
+	public String getCountryOfBirth() {
+		return this.countryOfBirth;
+	}
+
+	public void setCountryOfBirth(String countryOfBirth) {
+		this.countryOfBirth = countryOfBirth;
+	}
+
+}
