+++ src/main/java/co/edu/icesi/researchgroupmanagement/model/RgroupLeadersId.java
@@ -1,0 +1,64 @@
+package co.edu.icesi.researchgroupmanagement.model;
+// Generated Oct 22, 2022, 7:23:43 PM by Hibernate Tools 5.6.7.Final
+
+/**
+ * RgroupLeadersId generated by hbm2java
+ */
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
+@Entity
+public class RgroupLeadersId implements java.io.Serializable {
+
+	private long rgroupId;
+	private long usersId;
+
+	public RgroupLeadersId() {
+	}
+
+	public RgroupLeadersId(long rgroupId, long usersId) {
+		this.rgroupId = rgroupId;
+		this.usersId = usersId;
+	}
+
+	public long getRgroupId() {
+		return this.rgroupId;
+	}
+
+	@Id
+	@GeneratedValue(strategy= GenerationType.SEQUENCE)
+	public void setRgroupId(long rgroupId) {
+		this.rgroupId = rgroupId;
+	}
+
+	public long getUsersId() {
+		return this.usersId;
+	}
+
+	public void setUsersId(long usersId) {
+		this.usersId = usersId;
+	}
+
+	public boolean equals(Object other) {
+		if ((this == other))
+			return true;
+		if ((other == null))
+			return false;
+		if (!(other instanceof RgroupLeadersId))
+			return false;
+		RgroupLeadersId castOther = (RgroupLeadersId) other;
+
+		return (this.getRgroupId() == castOther.getRgroupId()) && (this.getUsersId() == castOther.getUsersId());
+	}
+
+	public int hashCode() {
+		int result = 17;
+
+		result = 37 * result + (int) this.getRgroupId();
+		result = 37 * result + (int) this.getUsersId();
+		return result;
+	}
+
+}
