+++ src/main/java/co/edu/icesi/researchgroupmanagement/model/RgroupUsers.java
@@ -1,17 +1,16 @@
 package co.edu.icesi.researchgroupmanagement.model;
-
-// Generated Oct 15, 2022, 11:56:22 PM by Hibernate Tools 5.6.7.Final
+// Generated Oct 22, 2022, 7:23:43 PM by Hibernate Tools 5.6.7.Final
 
 import java.util.Date;
 
+/**
+ * RgroupUsers generated by hbm2java
+ */
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 
-/**
- * RgroupUsers generated by hbm2java
- */
 @Entity
 public class RgroupUsers implements java.io.Serializable {
