+++ src/main/java/co/edu/icesi/researchgroupmanagement/model/RgroupUsersId.java
@@ -1,20 +1,17 @@
 package co.edu.icesi.researchgroupmanagement.model;
+// Generated Oct 22, 2022, 7:23:43 PM by Hibernate Tools 5.6.7.Final
 
+/**
+ * RgroupUsersId generated by hbm2java
+ */
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 
-// Generated Oct 15, 2022, 11:56:22 PM by Hibernate Tools 5.6.7.Final
-
-/**
- * RgroupUsersId generated by hbm2java
- */
 @Entity
 public class RgroupUsersId implements java.io.Serializable {
 
-	@Id
-	@GeneratedValue(strategy= GenerationType.SEQUENCE)
 	private long researchGroupId;
 	private long usersId;
 
@@ -30,6 +27,8 @@ public class RgroupUsersId implements java.io.Serializable {
 		return this.researchGroupId;
 	}
 
+	@Id
+	@GeneratedValue(strategy= GenerationType.SEQUENCE)
 	public void setResearchGroupId(long researchGroupId) {
 		this.researchGroupId = researchGroupId;
 	}
