+++ src/main/java/co/edu/icesi/researchgroupmanagement/model/RprojectRpapersId.java
@@ -1,20 +1,17 @@
 package co.edu.icesi.researchgroupmanagement.model;
+// Generated Oct 22, 2022, 7:23:43 PM by Hibernate Tools 5.6.7.Final
 
+/**
+ * RprojectRpapersId generated by hbm2java
+ */
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 
-// Generated Oct 15, 2022, 11:56:22 PM by Hibernate Tools 5.6.7.Final
-
-/**
- * RprojectRpapersId generated by hbm2java
- */
 @Entity
 public class RprojectRpapersId implements java.io.Serializable {
 
-	@Id
-	@GeneratedValue(strategy= GenerationType.SEQUENCE)
 	private long researchProjectId;
 	private long researchPaperId;
 
@@ -30,6 +27,8 @@ public class RprojectRpapersId implements java.io.Serializable {
 		return this.researchProjectId;
 	}
 
+	@Id
+	@GeneratedValue(strategy= GenerationType.SEQUENCE)
 	public void setResearchProjectId(long researchProjectId) {
 		this.researchProjectId = researchProjectId;
 	}
