+++ src/main/java/co/edu/icesi/researchgroupmanagement/model/RprojectStatus.java
@@ -1,39 +1,38 @@
 package co.edu.icesi.researchgroupmanagement.model;
-
-// Generated Oct 15, 2022, 11:56:22 PM by Hibernate Tools 5.6.7.Final
+// Generated Oct 22, 2022, 7:23:43 PM by Hibernate Tools 5.6.7.Final
 
 import java.util.HashSet;
 import java.util.Set;
 
+/**
+ * RprojectStatus generated by hbm2java
+ */
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 
-/**
- * AdmissionStatus generated by hbm2java
- */
 @Entity
-public class AdmissionStatus implements java.io.Serializable {
+public class RprojectStatus implements java.io.Serializable {
 
 	@Id
 	@GeneratedValue(strategy= GenerationType.SEQUENCE)
 	private long id;
 	private String status;
-	private Set admissionRequests = new HashSet(0);
+	private Set researchProjects = new HashSet(0);
 
-	public AdmissionStatus() {
+	public RprojectStatus() {
 	}
 
-	public AdmissionStatus(long id, String status) {
+	public RprojectStatus(long id, String status) {
 		this.id = id;
 		this.status = status;
 	}
 
-	public AdmissionStatus(long id, String status, Set admissionRequests) {
+	public RprojectStatus(long id, String status, Set researchProjects) {
 		this.id = id;
 		this.status = status;
-		this.admissionRequests = admissionRequests;
+		this.researchProjects = researchProjects;
 	}
 
 	public long getId() {
@@ -52,12 +51,12 @@ public class AdmissionStatus implements java.io.Serializable {
 		this.status = status;
 	}
 
-	public Set getAdmissionRequests() {
-		return this.admissionRequests;
+	public Set getResearchProjects() {
+		return this.researchProjects;
 	}
 
-	public void setAdmissionRequests(Set admissionRequests) {
-		this.admissionRequests = admissionRequests;
+	public void setResearchProjects(Set researchProjects) {
+		this.researchProjects = researchProjects;
 	}
 
 }
