+++ src/main/java/co/edu/icesi/researchgroupmanagement/model/RprojectTypes.java
@@ -1,38 +1,37 @@
 package co.edu.icesi.researchgroupmanagement.model;
-
-// Generated Oct 15, 2022, 11:56:22 PM by Hibernate Tools 5.6.7.Final
+// Generated Oct 22, 2022, 7:23:43 PM by Hibernate Tools 5.6.7.Final
 
 import java.util.HashSet;
 import java.util.Set;
 
+/**
+ * RprojectTypes generated by hbm2java
+ */
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 
-/**
- * FinancingSource generated by hbm2java
- */
 @Entity
-public class FinancingSource implements java.io.Serializable {
+public class RprojectTypes implements java.io.Serializable {
 
 	@Id
 	@GeneratedValue(strategy= GenerationType.SEQUENCE)
 	private long id;
-	private String source;
+	private String type;
 	private Set researchProjects = new HashSet(0);
 
-	public FinancingSource() {
+	public RprojectTypes() {
 	}
 
-	public FinancingSource(long id, String source) {
+	public RprojectTypes(long id, String type) {
 		this.id = id;
-		this.source = source;
+		this.type = type;
 	}
 
-	public FinancingSource(long id, String source, Set researchProjects) {
+	public RprojectTypes(long id, String type, Set researchProjects) {
 		this.id = id;
-		this.source = source;
+		this.type = type;
 		this.researchProjects = researchProjects;
 	}
 
@@ -44,12 +43,12 @@ public class FinancingSource implements java.io.Serializable {
 		this.id = id;
 	}
 
-	public String getSource() {
-		return this.source;
+	public String getType() {
+		return this.type;
 	}
 
-	public void setSource(String source) {
-		this.source = source;
+	public void setType(String type) {
+		this.type = type;
 	}
 
 	public Set getResearchProjects() {
