+++ src/main/java/co/edu/icesi/researchgroupmanagement/model/RprojectUsers.java
@@ -1,15 +1,14 @@
 package co.edu.icesi.researchgroupmanagement.model;
+// Generated Oct 22, 2022, 7:23:43 PM by Hibernate Tools 5.6.7.Final
 
+/**
+ * RprojectUsers generated by hbm2java
+ */
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 
-// Generated Oct 15, 2022, 11:56:22 PM by Hibernate Tools 5.6.7.Final
-
-/**
- * RprojectUsers generated by hbm2java
- */
 @Entity
 public class RprojectUsers implements java.io.Serializable {
