+++ src/main/java/co/edu/icesi/researchgroupmanagement/model/SciTiNationalProgram.java
@@ -1,44 +1,41 @@
 package co.edu.icesi.researchgroupmanagement.model;
-
-// Generated Oct 15, 2022, 11:56:22 PM by Hibernate Tools 5.6.7.Final
+// Generated Oct 22, 2022, 7:23:43 PM by Hibernate Tools 5.6.7.Final
 
 import java.util.HashSet;
 import java.util.Set;
 
+/**
+ * SciTiNationalProgram generated by hbm2java
+ */
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 
-/**
- * ResearchArea generated by hbm2java
- */
 @Entity
-public class ResearchArea implements java.io.Serializable {
+public class SciTiNationalProgram implements java.io.Serializable {
 
 	@Id
 	@GeneratedValue(strategy= GenerationType.SEQUENCE)
 	private long id;
-	private GreatRarea greatRarea;
 	private String name;
 	private String description;
-	private Set researchLines = new HashSet(0);
+	private Set researchGroups = new HashSet(0);
 
-	public ResearchArea() {
+	public SciTiNationalProgram() {
 	}
 
-	public ResearchArea(long id, String name, String description) {
+	public SciTiNationalProgram(long id, String name, String description) {
 		this.id = id;
 		this.name = name;
 		this.description = description;
 	}
 
-	public ResearchArea(long id, GreatRarea greatRarea, String name, String description, Set researchLines) {
+	public SciTiNationalProgram(long id, String name, String description, Set researchGroups) {
 		this.id = id;
-		this.greatRarea = greatRarea;
 		this.name = name;
 		this.description = description;
-		this.researchLines = researchLines;
+		this.researchGroups = researchGroups;
 	}
 
 	public long getId() {
@@ -47,14 +44,6 @@ public class ResearchArea implements java.io.Serializable {
 
 	public void setId(long id) {
 		this.id = id;
-	}
-
-	public GreatRarea getGreatRarea() {
-		return this.greatRarea;
-	}
-
-	public void setGreatRarea(GreatRarea greatRarea) {
-		this.greatRarea = greatRarea;
 	}
 
 	public String getName() {
@@ -73,12 +62,12 @@ public class ResearchArea implements java.io.Serializable {
 		this.description = description;
 	}
 
-	public Set getResearchLines() {
-		return this.researchLines;
+	public Set getResearchGroups() {
+		return this.researchGroups;
 	}
 
-	public void setResearchLines(Set researchLines) {
-		this.researchLines = researchLines;
+	public void setResearchGroups(Set researchGroups) {
+		this.researchGroups = researchGroups;
 	}
 
 }
