+++ src/main/java/co/edu/icesi/researchgroupmanagement/model/StudentProgram.java
@@ -1,18 +1,17 @@
 package co.edu.icesi.researchgroupmanagement.model;
-
-// Generated Oct 15, 2022, 11:56:22 PM by Hibernate Tools 5.6.7.Final
+// Generated Oct 22, 2022, 7:23:43 PM by Hibernate Tools 5.6.7.Final
 
 import java.util.HashSet;
 import java.util.Set;
 
+/**
+ * StudentProgram generated by hbm2java
+ */
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 
-/**
- * StudentProgram generated by hbm2java
- */
 @Entity
 public class StudentProgram implements java.io.Serializable {
