+++ src/main/java/co/edu/icesi/researchgroupmanagement/model/UserTypes.java
@@ -1,39 +1,38 @@
 package co.edu.icesi.researchgroupmanagement.model;
-
-// Generated Oct 15, 2022, 11:56:22 PM by Hibernate Tools 5.6.7.Final
+// Generated Oct 22, 2022, 7:23:43 PM by Hibernate Tools 5.6.7.Final
 
 import java.util.HashSet;
 import java.util.Set;
 
+/**
+ * UserTypes generated by hbm2java
+ */
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 
-/**
- * FinancingType generated by hbm2java
- */
 @Entity
-public class FinancingType implements java.io.Serializable {
+public class UserTypes implements java.io.Serializable {
 
 	@Id
 	@GeneratedValue(strategy= GenerationType.SEQUENCE)
 	private long id;
 	private String type;
-	private Set researchProjects = new HashSet(0);
+	private Set userses = new HashSet(0);
 
-	public FinancingType() {
+	public UserTypes() {
 	}
 
-	public FinancingType(long id, String type) {
+	public UserTypes(long id, String type) {
 		this.id = id;
 		this.type = type;
 	}
 
-	public FinancingType(long id, String type, Set researchProjects) {
+	public UserTypes(long id, String type, Set userses) {
 		this.id = id;
 		this.type = type;
-		this.researchProjects = researchProjects;
+		this.userses = userses;
 	}
 
 	public long getId() {
@@ -52,12 +51,12 @@ public class FinancingType implements java.io.Serializable {
 		this.type = type;
 	}
 
-	public Set getResearchProjects() {
-		return this.researchProjects;
+	public Set getUserses() {
+		return this.userses;
 	}
 
-	public void setResearchProjects(Set researchProjects) {
-		this.researchProjects = researchProjects;
+	public void setUserses(Set userses) {
+		this.userses = userses;
 	}
 
 }
