+++ src/main/java/co/edu/icesi/researchgroupmanagement/model/Users.java
@@ -1,19 +1,18 @@
 package co.edu.icesi.researchgroupmanagement.model;
-
-// Generated Oct 15, 2022, 11:56:22 PM by Hibernate Tools 5.6.7.Final
+// Generated Oct 22, 2022, 7:23:43 PM by Hibernate Tools 5.6.7.Final
 
 import java.util.Date;
 import java.util.HashSet;
 import java.util.Set;
 
+/**
+ * Users generated by hbm2java
+ */
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 
-/**
- * Users generated by hbm2java
- */
 @Entity
 public class Users implements java.io.Serializable {
 
@@ -29,7 +28,7 @@ public class Users implements java.io.Serializable {
 	private String email;
 	private byte[] userPicture;
 	private Set admissionRequests = new HashSet(0);
-	private Set researchGroups = new HashSet(0);
+	private Set rgroupLeaderses = new HashSet(0);
 	private Set rgroupUserses = new HashSet(0);
 	private Set studentPrograms = new HashSet(0);
 	private Set resources = new HashSet(0);
@@ -49,7 +48,7 @@ public class Users implements java.io.Serializable {
 	}
 
 	public Users(long id, UserTypes userTypes, String username, String firstName, String lastName, Date dob,
-			String password, String email, byte[] userPicture, Set admissionRequests, Set researchGroups,
+			String password, String email, byte[] userPicture, Set admissionRequests, Set rgroupLeaderses,
 			Set rgroupUserses, Set studentPrograms, Set resources, Set rprojectUserses) {
 		this.id = id;
 		this.userTypes = userTypes;
@@ -61,7 +60,7 @@ public class Users implements java.io.Serializable {
 		this.email = email;
 		this.userPicture = userPicture;
 		this.admissionRequests = admissionRequests;
-		this.researchGroups = researchGroups;
+		this.rgroupLeaderses = rgroupLeaderses;
 		this.rgroupUserses = rgroupUserses;
 		this.studentPrograms = studentPrograms;
 		this.resources = resources;
@@ -148,12 +147,12 @@ public class Users implements java.io.Serializable {
 		this.admissionRequests = admissionRequests;
 	}
 
-	public Set getResearchGroups() {
-		return this.researchGroups;
+	public Set getRgroupLeaderses() {
+		return this.rgroupLeaderses;
 	}
 
-	public void setResearchGroups(Set researchGroups) {
-		this.researchGroups = researchGroups;
+	public void setRgroupLeaderses(Set rgroupLeaderses) {
+		this.rgroupLeaderses = rgroupLeaderses;
 	}
 
 	public Set getRgroupUserses() {
